Do not mutate receiver in ToWire call
Summary:
Avoid mutating the receiver with default values in `ToWire` calls
by changing the template to use a local variable to store defaults
rather than writing to the object. Also, fix tests accoringly.

*Note*: this change should be rolled out with caution, because,
as many of our own tests depended on this behavior, it is possible
that other users might be broken by this change.